% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_posthoc_test.R
\name{chisq.posthoc.test}
\alias{chisq.posthoc.test}
\title{Perform post hoc analysis based on residuals of Pearson's Chi-squared Test for Count Data.}
\usage{
chisq.posthoc.test(x, y, method = "bonferroni", round = 6, ...)
}
\arguments{
\item{x}{A numeric vector passed on to the chisq.test function.}

\item{y}{A numeric vector passed on to the chisq.test function.}

\item{method}{The p adjustment method to be used. This is passed on to the p.adjust function.}

\item{round}{Number of digits to round the p.value to. Defaults to 6.}

\item{...}{Additional arguments passed on to the chisq.test function.}
}
\value{
A table with the adjusted p value for each x y combination.
}
\description{
Perform post hoc analysis based on residuals of Pearson's Chi-squared Test for Count Data.
}
